contoso-hr-agent/pyproject.toml
[project]
name = "contoso-hr"
version = "0.1.0"
description = "Contoso HR Agent: AI-powered resume screening and HR policy Q&A"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "O'Reilly Training" }]
keywords = ["ai", "agents", "langgraph", "crewai", "mcp", "hr", "azure"]

dependencies = [
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "crewai>=0.80.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "fastmcp>=2.0.0,<3.0.0",
    "chromadb>=0.5.0",
    "httpx>=0.27.0",
    "requests>=2.31.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[project.scripts]
hr-engine   = "contoso_hr.engine:main"
hr-watcher  = "contoso_hr.watcher.resume_watcher:main"
hr-mcp      = "contoso_hr.mcp_server.__main__:main"
hr-seed     = "contoso_hr.knowledge.vectorizer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/contoso_hr"]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
